movie_review_project/requirements.txt
fastapi
tortoise-orm
pydantic
uvicorn
python-dotenv

